# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Searches the configured public reservation source for sailings matching the requested filters.
 * @summary Search available boat fishing schedules
 */
export const SearchFishingSchedulesQueryParams = zod.object({
  "startDate": zod.date(),
  "endDate": zod.date(),
  "region": zod.array(zod.coerce.string()).optional(),
  "port": zod.coerce.string().optional(),
  "ship": zod.array(zod.coerce.string()).optional(),
  "tide": zod.coerce.string().optional()
})

export const SearchFishingSchedulesResponse = zod.object({
  "items": zod.array(zod.object({
  "id": zod.string(),
  "departureDate": zod.coerce.date(),
  "weekday": zod.string(),
  "region": zod.string(),
  "port": zod.string(),
  "tide": zod.string(),
  "tideStrengthPercent": zod.number().int().nullable().describe('바다타임(badatime.com) 기준 조류 세기(%). 매핑되지 않은 항구이거나 조회에 실패하면 null.'),
  "genre": zod.string(),
  "operator": zod.string(),
  "vessel": zod.string(),
  "remainingSeats": zod.number().int().nullable(),
  "operatorUrl": zod.string().url().nullable(),
  "bookingUrl": zod.string().url().nullable(),
  "source": zod.string()
})),
  "total": zod.number().int(),
  "searchedAt": zod.coerce.date(),
  "source": zod.string().describe('URL for one source, or a label when multiple sources are configured'),
  "cachedUntil": zod.coerce.date().nullable(),
  "warning": zod.string().nullable()
})


/**
 * Returns filter values discovered from the configured public reservation source.
 * @summary Get available fishing filters
 */
export const GetFishingFilterOptionsResponse = zod.object({
  "regions": zod.array(zod.string()),
  "ports": zod.array(zod.string()),
  "ships": zod.array(zod.string()),
  "tides": zod.array(zod.string()),
  "shipsByPort": zod.record(zod.string(), zod.array(zod.string())),
  "shipsByRegion": zod.record(zod.string(), zod.array(zod.string()))
})


/**
 * @summary List configured fishing reservation sources
 */
export const ListFishingSourcesResponseItem = zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "region": zod.string(),
  "port": zod.string(),
  "sourceUrl": zod.string().url(),
  "enabled": zod.boolean(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date(),
  "vessels": zod.array(zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "departurePort": zod.string(),
  "address": zod.string()
}))
})
export const ListFishingSourcesResponse = zod.array(ListFishingSourcesResponseItem)


/**
 * @summary Add a fishing reservation source
 */



export const createFishingSourceBodyEnabledDefault = true;

export const CreateFishingSourceBody = zod.object({
  "name": zod.string().min(1),
  "region": zod.string().min(1),
  "port": zod.string().min(1),
  "sourceUrl": zod.string().url(),
  "enabled": zod.boolean().default(createFishingSourceBodyEnabledDefault)
})

export const CreateFishingSourceResponse = zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "region": zod.string(),
  "port": zod.string(),
  "sourceUrl": zod.string().url(),
  "enabled": zod.boolean(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date(),
  "vessels": zod.array(zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "departurePort": zod.string(),
  "address": zod.string()
}))
})


/**
 * @summary Update a fishing reservation source
 */
export const UpdateFishingSourceParams = zod.object({
  "id": zod.coerce.number().int()
})




export const updateFishingSourceBodyEnabledDefault = true;

export const UpdateFishingSourceBody = zod.object({
  "name": zod.string().min(1),
  "region": zod.string().min(1),
  "port": zod.string().min(1),
  "sourceUrl": zod.string().url(),
  "enabled": zod.boolean().default(updateFishingSourceBodyEnabledDefault)
})

export const UpdateFishingSourceResponse = zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "region": zod.string(),
  "port": zod.string(),
  "sourceUrl": zod.string().url(),
  "enabled": zod.boolean(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date(),
  "vessels": zod.array(zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "departurePort": zod.string(),
  "address": zod.string()
}))
})


/**
 * @summary Delete a fishing reservation source
 */
export const DeleteFishingSourceParams = zod.object({
  "id": zod.coerce.number().int()
})

export const DeleteFishingSourceResponse = zod.void()


/**
 * @summary Add vessels to a fishing reservation source
 */
export const AddFishingSourceVesselsParams = zod.object({
  "id": zod.coerce.number().int()
})





export const AddFishingSourceVesselsBody = zod.object({
  "vessels": zod.array(zod.object({
  "name": zod.string().min(1),
  "departurePort": zod.string(),
  "address": zod.string()
})).min(1)
})

export const AddFishingSourceVesselsResponse = zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "region": zod.string(),
  "port": zod.string(),
  "sourceUrl": zod.string().url(),
  "enabled": zod.boolean(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date(),
  "vessels": zod.array(zod.object({
  "id": zod.number().int(),
  "name": zod.string(),
  "departurePort": zod.string(),
  "address": zod.string()
}))
})


/**
 * @summary Clear manually registered vessels for a source
 */
export const ClearFishingSourceVesselsParams = zod.object({
  "id": zod.coerce.number().int()
})

export const ClearFishingSourceVesselsResponse = zod.void()


/**
 * @summary Probe a URL to guess its reservation platform, name, and vessel list before registering it as a source
 */
export const DetectFishingSourceBody = zod.object({
  "url": zod.string()
})

export const DetectFishingSourceResponse = zod.object({
  "platform": zod.string().describe('"thefishing" | "sunsang24" | "unknown" 중 하나. 더피싱(index.php?mid=bk) 또는 SUNSANG24(/ship/schedule_fleet) 패턴 중 실제로 선박이 긁힌 쪽이고, 둘 다 안 되면 unknown.'),
  "suggestedName": zod.string().nullable().describe('사이트 <title> 태그에서 뽑아본 이름 추천값. 참고용이라 그대로 안 써도 됨.'),
  "vessels": zod.array(zod.string()).describe('이번 달(또는 비어있으면 다음 달) 조회에서 실제로 발견된 선박 이름들.'),
  "workingUrl": zod.string().nullable().describe('정상 인식된 경우의 기준 URL. platform이 unknown이면 null.')
})