# lib/api-zod/src/generated/types/detectFishingSourceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectFishingSourceResponse {
  /** "thefishing" | "sunsang24" | "unknown" 중 하나. 더피싱(index.php?mid=bk) 또는 SUNSANG24(/ship/schedule_fleet) 패턴 중 실제로 선박이 긁힌 쪽이고, 둘 다 안 되면 unknown. */
  platform: string;
  /**
     * 사이트 <title> 태그에서 뽑아본 이름 추천값. 참고용이라 그대로 안 써도 됨.
     * @nullable
     */
  suggestedName: string | null;
  /** 이번 달(또는 비어있으면 다음 달) 조회에서 실제로 발견된 선박 이름들. */
  vessels: string[];
  /**
     * 정상 인식된 경우의 기준 URL. platform이 unknown이면 null.
     * @nullable
     */
  workingUrl: string | null;
}